[WCNCR00303306] Fix coverity
Coverity issue from
CE-neptune-main_MT6639_X86_PCIE_MOBILE_GEN4M_COV_20230425041002

Issue types:
1. UNINIT
2. UNUSED_VALUE
3. DEADCODE
4. REVERSE_INULL
5. PRINTF_ARGS
6. CERT EXP34-C/CERT STR31-C
7. FORWARD_NULL
8. RESOURCE_LEAK
9. OVERRUN

MTK-Commit-Id: f481c2c7f07c193e6ed190e8e2bf7d0c49682065

CR-Id: WCNCR00303306
Feature: Wi-Fi Driver
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/7426462
Test: srv_pf_nep_sanity
Build: srv_pf_nep_win <srv_pf_nep_win@mediatek.com>
Build: srv_preflight_b001 <srv_preflight_b001@mediatek.com>
Coverity-Review-Label: srv_pf_nep_win <srv_pf_nep_win@mediatek.com>
Commit-Check: srv_pis <srv_pis@mediatek.com>
Coverity-Review-Label: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Test: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Test: srv_mspautosanity

+++ include/mgmt/fils.h
@@ -65,8 +65,8 @@
 
 struct EAP_ERP_KEY {
 	uint8_t fgValid;
-	uint8_t rRKLen;
-	uint8_t rIKLen;
+	uint16_t rRKLen;
+	uint16_t rIKLen;
 	uint8_t rRK[ERP_MAX_KEY_LEN];
 	uint8_t rIK[ERP_MAX_KEY_LEN];
 	uint32_t nextSeq; /* next seq */
